Set parameters for graph optimisation

There are two steps needed to fine tune the graph optimisation with multigrain:

Create a control object with multigrain_control().

Define its behaviour with control_ functions:

control_nsim_local() to set the number of local simulations.

control_nsim_global() to set the number of global simulations.

control_local() to modify the local optimisation options.

control_global() to modify the global optimisation options.

Any unset parameters will automatically be set before running the optimisation. There are predefined defaults, but they are calibrated based on the pvals dimensions.

Usage

multigrain_control()

Value

A multigrain_control object.

Examples

multigrain_control()
#> <multigrain_control>